Set up project to automatically run in qemu
Also add simple "Hello World!" functionality to `_start` in order to
test the result. It uses the VGA buffer which is compatible with a huge
range of machines.

So far we had a standalone executable but it wasn't really bootable. We
use the `bootloader` crate and the `bootimage` tool to create a bootable
image. We also set up `cargo run` to run the produced image through qemu
for easy testing.

Note that apparently, newer versions of `bootloader` do not require
`bootimage`, but I don't want to mess with stuff that is not in the
tutorial yet. It is a good idea for a future improvement after I'm done
with the main part of the tutorial.